Reduce dockerfile layers
The apt-get call now follows the docker best pratices from
https://docs.docker.com/develop/develop-images/dockerfile_best-practices/#run
and reduces the layer size.

COPY is now used instead of ADD as the former can automatically extract
archives, which we don't need. `download.sh` is now explicitly specified
and the chmod&execution is now a single RUN command.

Similarly the user and build directory setup is now a single command.